requirements.txt
aioredis
bluesky
bluesky-kafka
databroker
fastapi
json-rpc
jsonschema
matplotlib # Needed for BEC, should be factored out
scikit-image # Needed for BEC/other stuff to run, factor out
msgpack>=1.0.0
msgpack_numpy
ophyd
pydantic
pyyaml
pyzmq
requests
uvicorn
